Fixed typs and floats and ints and penises

+++ TrimmDich/src/de/lmu/msp/trimmdich/dialog/NumberPickerFragement.java
@@ -91,7 +91,7 @@ public class NumberPickerFragement extends DialogFragment {
 			mSPrefs = getActivity().getSharedPreferences(Constants.PREFS_NAME,
 					0);
 			SharedPreferences.Editor editor = mSPrefs.edit();
-			editor.putFloat(mPickerSPref, (float) mNumberPicker.getValue());
+			editor.putInt(mPickerSPref, mNumberPicker.getValue());
 			editor.commit();
 
 			((MainActivity) getActivity()).doPositiveClick();
